fix(contracts): change type of gasLimit in PublicInput to uint64
Due to the discrepancy of the type of gasLimit in PublicInput,
the different hash of PublicInput was returned.
Thus, the type has been changed to uint64

+++ packages/contracts/contracts/libraries/Types.sol
@@ -60,7 +60,7 @@ library Types {
         bytes32 parentHash;
         uint64 timestamp;
         uint64 number;
-        uint256 gasLimit;
+        uint64 gasLimit;
         uint256 baseFee;
         bytes32 transactionsRoot;
         bytes32 stateRoot;
